+++ crates/nostr/src/nips/nip34.rs
@@ -16,15 +16,14 @@ use core::fmt;
 use hashes::sha1::Hash as Sha1Hash;
 
 use crate::event::builder::{Error, EventBuilder, WrongKindError};
-use crate::nips::nip01::Coordinate;
+use crate::nips::nip01::{self, Coordinate};
 use crate::nips::nip10::Marker;
 use crate::types::url::Url;
 use crate::{EventId, Kind, PublicKey, RelayUrl, Tag, TagKind, TagStandard, Timestamp};
 
-/// Earlier unique commit ID
+/// Earlier unique commit ID marker
 pub const EUC: &str = "euc";
 
-const GIT_REPO_ANNOUNCEMENT_ALT: &str = "git repository";
 const GIT_PATCH_ALT: &str = "git patch";
 const GIT_PATCH_COVER_LETTER_ALT: &str = "git patch cover letter";
 
@@ -52,13 +51,18 @@ pub struct GitRepositoryAnnouncement {
     /// made to identify it among forks and group it with other repositories hosted elsewhere that may represent essentially the same project.
     /// In most cases it will be the root commit of a repository.
     /// In case of a permanent fork between two projects, then the first commit after the fork should be used.
-    pub euc: Option<String>,
+    pub euc: Option<Sha1Hash>,
     /// Other recognized maintainers
     pub maintainers: Vec<PublicKey>,
 }
 
 impl GitRepositoryAnnouncement {
-    pub(crate) fn to_event_builder(self) -> EventBuilder {
+    pub(crate) fn to_event_builder(self) -> Result<EventBuilder, Error> {
+        if self.id.is_empty() {
+            // TODO: should return another error?
+            return Err(Error::NIP01(nip01::Error::InvalidCoordinate));
+        }
+
         let mut tags: Vec<Tag> = Vec::with_capacity(1);
 
         // Add repo ID
@@ -98,9 +102,9 @@ impl GitRepositoryAnnouncement {
         }
 
         // Add EUC
-        if let Some(euc) = self.euc {
+        if let Some(commit) = self.euc {
             tags.push(Tag::from_standardized_without_cell(
-                TagStandard::GitEarliestUniqueCommitId(euc),
+                TagStandard::GitEarliestUniqueCommitId(commit),
             ));
         }
 
@@ -111,11 +115,8 @@ impl GitRepositoryAnnouncement {
             ));
         }
 
-        // Add alt tag
-        tags.push(Tag::alt(GIT_REPO_ANNOUNCEMENT_ALT));
-
         // Build
-        EventBuilder::new(Kind::GitRepoAnnouncement, "").tags(tags)
+        Ok(EventBuilder::new(Kind::GitRepoAnnouncement, "").tags(tags))
     }
 }
 
@@ -330,8 +331,55 @@ impl GitPatch {
 
 #[cfg(all(test, feature = "std"))]
 mod tests {
+    use core::str::FromStr;
+
     use super::*;
     use crate::{Event, Keys, Tags};
+
+    #[test]
+    fn test_git_repo_announcement() {
+        let repo = GitRepositoryAnnouncement {
+            id: String::from("test"),
+            name: Some(String::from("Test nostr repository")),
+            description: Some(String::from("Long desc")),
+            web: Vec::new(),
+            clone: vec![Url::parse("https://github.com/rust-nostr/nostr.git").unwrap()],
+            relays: vec![
+                RelayUrl::parse("wss://example.com").unwrap(),
+                RelayUrl::parse("wss://example.org").unwrap(),
+            ],
+            euc: Some(Sha1Hash::from_str("aa231c4c6a5777dc89b42207b499891a344add5c").unwrap()),
+            maintainers: vec![PublicKey::parse(
+                "npub1drvpzev3syqt0kjrls50050uzf25gehpz9vgdw08hvex7e0vgfeq0eseet",
+            )
+            .unwrap()],
+        };
+
+        let keys = Keys::generate();
+        let event: Event = repo
+            .to_event_builder()
+            .unwrap()
+            .sign_with_keys(&keys)
+            .unwrap();
+
+        assert_eq!(event.kind, Kind::GitRepoAnnouncement);
+        assert!(event.content.is_empty());
+
+        let tags = Tags::parse([
+            vec!["d", "test"],
+            vec!["name", "Test nostr repository"],
+            vec!["description", "Long desc"],
+            vec!["clone", "https://github.com/rust-nostr/nostr.git"],
+            vec!["relays", "wss://example.com", "wss://example.org"],
+            vec!["r", "aa231c4c6a5777dc89b42207b499891a344add5c", "euc"],
+            vec![
+                "maintainers",
+                "68d81165918100b7da43fc28f7d1fc12554466e1115886b9e7bb326f65ec4272",
+            ],
+        ])
+        .unwrap();
+        assert_eq!(event.tags, tags);
+    }
 
     #[test]
     fn test_git_issue() {